Move provider-dev manager into providerdev service
Move provider-dev session and attachment state management out of gestaltd/internal/providerdev and into services/providerdev, next to provider-dev HTTP route wiring.

Interface diff:

- import "github.com/valon-technologies/gestalt/server/internal/providerdev"

- providerdev.NewManager(targets, opts...)

+ import "github.com/valon-technologies/gestalt/server/services/providerdev"

+ providerdev.NewManager(targets, opts...)

Example:

manager, err := providerdev.NewManager(targets, providerdev.WithIndexedDBAttachmentState(ctx, db))
