+++ gestaltd/internal/bootstrap/cleanup.go
@@ -6,8 +6,8 @@ import (
 	"io"
 
 	"github.com/valon-technologies/gestalt/server/core"
-	"github.com/valon-technologies/gestalt/server/internal/providerdev"
 	"github.com/valon-technologies/gestalt/server/services/plugins/registry"
+	"github.com/valon-technologies/gestalt/server/services/providerdev"
 )
 
 // CloseProviders closes all registered providers that implement io.Closer.
