+++ gestaltd/internal/server/handlers_provider_dev.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/go-chi/chi/v5"
 	"github.com/valon-technologies/gestalt/server/core"
-	"github.com/valon-technologies/gestalt/server/internal/providerdev"
 	"github.com/valon-technologies/gestalt/server/services/identity/principal"
+	"github.com/valon-technologies/gestalt/server/services/providerdev"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
